feat: Add delta noise adjustment to latent model input
This commit adds a delta noise adjustment to the latent model input in the cogvideox_sampler function. The added noise is concatenated with the latent model input along the third dimension. This change allows for later rectification of the added noise. The image_rotary_emb variable is now set to the image_emb value for all iterations except the last one.

Ref: #issue-number